+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to [diagram-js](https://github.com/bpmn-io/diagram-js) are d
 _**Note:** Yet to be released changes appear here._
 
 * `FEAT`: allow `Canvas#findRoot` to find shared root for multiple elements ([#1075](https://github.com/bpmn-io/diagram-js/pull/1075))
+* `FEAT`: support scrolling to multiple elements via `Canvas#scrollToElement` ([#1075](https://github.com/bpmn-io/diagram-js/pull/1075))
 
 ## 15.19.0